feat(*): closing registrations
