+++ src-ebleague/app/tournaments/tournament-registration/tournament-registration.component.scss
@@ -1,0 +1,51 @@
+.registration-title {
+  margin-top: 8px;
+}
+
+.application-details {
+  display: flex;
+  flex-flow: column;
+  align-items: center;
+  height: 0;
+  overflow: hidden;
+}
+
+.application-details-show {
+  height: 308px;
+  transition: height 1s .4s;
+}
+
+.accept-terms-checkbox {
+  margin: 12px;
+}
+
+.submit-registration {
+  height: 48px;
+  margin: 8px auto;
+}
+
+.registration-form-input {
+  min-width: 380px;
+  margin: 4px 0;
+}
+
+.team-name {
+  padding: 0 8px;
+}
+
+.team-avatar {
+  width: 32px;
+  min-width: 32px;
+  height: 32px;
+}
+
+.success-dialog-content {
+  width: 550px;
+  padding: 16px;
+}
+
+@media only screen and (max-width: 550px) {
+  .success-dialog-content {
+    width: 380px;
+  }
+}
