+++ src-ebleague/app/tournaments/tournament-registration/tournament-registration.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { TournamentRegistrationComponent } from './tournament-registration.component';
+
+describe('TournamentRegistrationComponent', () => {
+  let component: TournamentRegistrationComponent;
+  let fixture: ComponentFixture<TournamentRegistrationComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ TournamentRegistrationComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(TournamentRegistrationComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
