+++ src-ebleague/styles.scss
@@ -170,7 +170,7 @@ $overlay-theme: igx-overlay-theme($background-color: rgba(35, 51, 132, 0.7));
 
 .registration-badge-style {
   position: absolute;
-  left: 80px;
+  left: 104px;
 }
 
 .empty-group-list {
